backend/requirements.txt
fastapi
uvicorn
langchain-openai
langgraph
langgraph-checkpoint-postgres
asyncpg
pydantic
python-dotenv